requirements.txt
argcomplete==1.11.1
asgiref==3.2.7
boto3==1.13.2
botocore==1.16.2
certifi==2020.4.5.1
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
Django==3.0.6
django-s3-storage>=0.13.0
djangorestframework==3.11.0
docutils==0.15.2
durationpy==0.5
future==0.18.2
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
kappa==0.6.0
pip-tools==5.1.2
placebo==0.9.0
psycopg2-binary==2.8.5
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
tqdm==4.46.0
troposphere==2.6.1
urllib3==1.25.9
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
